feat: updated ai service backend
Renamed ai-chat-service to gen-ai-service

Added two new endpoints:
- Endpoint for analysis of failed test cases
- Endpoint for analysis of error logs

Also added chat session management with in-memory storage for user-specific sessions using API key and room ID. Now, each session for each user will have a newly instantiated chatSession, so the history will only contain chats within that session. This ensures that the responses are not affected by previous chat history that may be unrelated, which was the case previously as only one chat session is ever instantiated.

Frontend is also to now pass the google AI api key to backend when making requests, instead of relying on the API key in the .env file